telegram-bot-hosh-gpt/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
pydantic==2.9.1
pydantic-settings==2.4.0
pydantic_core==2.23.3
python-dotenv==1.0.1
python-telegram-bot==21.5
redis==5.0.8
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
